Refactor xtask branding command into modular structure
